Issue #12 Regex changes
Confirmed working with Firefox 20 + Firebug & Chrome 26. Couldn't get
it to run properly on Webkit nightly.

Regex isn't escaping the ) parenthesis or the bracket which is why it
wouldn't match.  After thinking about it and submitting a req for an
altered regex I realized that for these functions they don't care
what's in the backreference we just want to style it all.